requirements.txt
Django==1.6
Pillow==2.6.1
South==1.0.1
django-compressor==1.4
django-debug-toolbar==1.2.2
django-annoying==0.8.0
